# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ReviewDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"game_id": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"review_text": {
						"type": "string"
					}
				},
				"required": [
					"game_id",
					"rating"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameOutputDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"console": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"title",
					"console"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameInputDTO": {
				"properties": {
					"title": {
						"type": "string"
					},
					"console_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"console_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_GameInputDTO_": {
				"properties": {
					"title": {
						"type": "string"
					},
					"console_id": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"ConsoleDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"manufacturer"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-review-game",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/reviews": {
			"get": {
				"operationId": "GetAllReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ReviewDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateReview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewDTO"
							}
						}
					}
				}
			}
		},
		"/reviews/{id}": {
			"get": {
				"operationId": "GetReviewById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ReviewDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ReviewDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteReview",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games": {
			"get": {
				"operationId": "GetAllGames",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GameOutputDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GameInputDTO"
							}
						}
					}
				}
			}
		},
		"/games/{id}": {
			"get": {
				"operationId": "GetGameById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/GameOutputDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/GameOutputDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_GameInputDTO_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteGame",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/games/{id}/reviews": {
			"get": {
				"operationId": "GetReviewsByGameId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ReviewDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/consoles": {
			"get": {
				"operationId": "GetAllConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ConsoleDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConsoleDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsoleDTO"
							}
						}
					}
				}
			}
		},
		"/consoles/{id}": {
			"get": {
				"operationId": "GetConsoleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ConsoleDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteConsole",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateConsole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ConsoleDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConsoleDTO"
							}
						}
					}
				}
			}
		},
		"/consoles/{id}/games": {
			"get": {
				"operationId": "GetGamesByConsoleId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GameOutputDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Consoles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}